Refactored move_node_directories, thanks Stan!
