+++ src/io.rs
@@ -1,7 +1,7 @@
 use serde_json::Value;
 use std::{
     path::{Path, PathBuf},
-    io::{self, Write},
+    io::Write,
     process::{Command, Stdio, exit},
     env,
     fs
